+++ src/xa51/ralloc.c
@@ -199,8 +199,8 @@ static void freeReg (regs * reg, bool silent) {
 /* allocReg - allocates register of given size (byte, word, dword) */
 /*            and type (ptr, gpr, cnd)                             */
 /*-----------------------------------------------------------------*/
-static bool allocReg (short size, short type, symbol *sym, 
-		      short offset, bool silent) {
+static bool allocReg (unsigned int size, int type, symbol *sym, 
+		      int offset, bool silent) {
   int i;
   
   checkRegMask(__FUNCTION__);
